Allow specifying target count
As of now, to launch the demo, two platforms are needed. This will
strike back during development and if there will be a need to integrate
solutions with other platforms.

This commit adds SINGLE_TARGET option, that when set will deploy client
certificates on server target. This way demo can be tested on a single
platform.

Signed-off-by: Mateusz Kusiak <mateusz.kusiak@3mdeb.com>